Move hasEmergencySymptoms method to symptomHistory
